Sensor values mapped to percentages (BACKEND)
ESP32 has 12-bit ADC so all analog values are in the range of 0 - 4096
I could have converted the voltage readings to the unit of measurement of the parameters
such as moisture is measured as g/m^3 i.e grams of water per cubic meter of air
and light intensity is measured in terms of candela. However, a layman user generally do not have the idea of these units
therefore, it is best if the values are mapped to percentages. E.g no water -> 0% | distilled water -> 100%
No light -> 0% | Complete light -> 100%