JSON parsing successful
DISCLAIMER: LONG ONE
So, module A can send a string which is basically in JSON format. However, this was not easy.
The datatype of JSON_data is String which is available in C++. The problem however is that, this cannot be written to
serial connection. Only char array i.e C type string is supported. Therefore I had to use .toCharArray to convert the String
to char[] and save it to a char[] buffer. Only then I could write it to serial.

On module B I used the ArduinoJSON.h file for JSON deserialization/parsing and extracted the data.