Integrated JSON parsing and Cloud data transfer to cloud.ino and updated wifi methods
Now, the device will automatically connect to the previous saved network. If the network is unavailable
it starts the config portal. If the Wifi goes off or the push button is pressed the config portal starts again.
If and only if the device is connected to a network AND data is available on serial2, the data will be read in form of
string which is in JSON format and will parse it and extract the sensor data and then upload it to the ThingSpeak cloud
using the API and Channel ID.