AP_STA enabled
Removed the requirement of module B by making module A work as both as an Access Point and Station SIMULTANOUSLY.
There is no more requirement of Serial Communication.
However, sometimes the mDNS server does not work properly.
The websocket function is now called after 17 seconds in place of 1 second.
Also, I had to replace WiFiManager with ESPAsyncWiFiManager as the former library was incompatible with TCPAsync library.